+++ etils/epy/adhoc_utils/binary_import.py
@@ -19,8 +19,6 @@ from __future__ import annotations
 from collections.abc import Iterator
 import contextlib
 import functools
-import getpass
-import os
 import sys
 from typing import Optional
 
@@ -67,6 +65,5 @@ def binary_adhoc(
     restrict: None | py_utils.StrOrStrList = None,
     reload: None | py_utils.StrOrStrList = None,
     verbose: bool = False,
-    legacy_backend: bool | None = None,
 ) -> Iterator[None]:
   yield